Move history panel specific JavaScript to its own file
Like the timer panel, the JavaScript that only applies to the history
panel should be contained within its own JavaScript module. This avoids
loading the history event handler unnecessarily.

The utility functions $$ and ajax have been moved to a utils.js es6
module so they can be imported by the main toolbar.js file as well as in
panel specific scripts.

The new utils.js module also helps move Django Debug Toolbar's
JavaScript files to a more modularized approached.
