+++ debug_toolbar/static/debug_toolbar/js/toolbar.js
@@ -1,76 +1,10 @@
-const $$ = {
-    on: function (root, eventName, selector, fn) {
-        root.addEventListener(eventName, function (event) {
-            const target = event.target.closest(selector);
-            if (root.contains(target)) {
-                fn.call(target, event);
-            }
-        });
-    },
-    show: function (element) {
-        element.classList.remove("djdt-hidden");
-    },
-    hide: function (element) {
-        element.classList.add("djdt-hidden");
-    },
-    toggle: function (element, value) {
-        if (value) {
-            $$.show(element);
-        } else {
-            $$.hide(element);
-        }
-    },
-    visible: function (element) {
-        element.classList.contains("djdt-hidden");
-    },
-    executeScripts: function (scripts) {
-        scripts.forEach(function (script) {
-            const el = document.createElement("script");
-            el.type = "module";
-            el.src = script;
-            el.async = true;
-            document.head.appendChild(el);
-        });
-    },
-};
+import { $$, ajax } from "./utils.js";
 
 const onKeyDown = function (event) {
     if (event.keyCode === 27) {
         djdt.hide_one_level();
     }
 };
-
-const ajax = function (url, init) {
-    init = Object.assign({ credentials: "same-origin" }, init);
-    return fetch(url, init).then(function (response) {
-        if (response.ok) {
-            return response.json();
-        } else {
-            const win = document.querySelector("#djDebugWindow");
-            win.innerHTML =
-                '<div class="djDebugPanelTitle"><a class="djDebugClose" href="">»</a><h3>' +
-                response.status +
-                ": " +
-                response.statusText +
-                "</h3></div>";
-            $$.show(win);
-            return Promise.reject();
-        }
-    });
-};
-
-function ajaxForm(element) {
-    const form = element.closest("form");
-    const url = new URL(form.action);
-    const formData = new FormData(form);
-    for (const [name, value] of formData.entries()) {
-        url.searchParams.append(name, value);
-    }
-    const ajaxData = {
-        method: form.method.toUpperCase(),
-    };
-    return ajax(url, ajaxData);
-}
 
 const djdt = {
     handleDragged: false,
@@ -154,50 +88,6 @@ const djdt = {
                 const win = djDebug.querySelector("#djDebugWindow");
                 win.innerHTML = data.content;
                 $$.show(win);
-            });
-        });
-
-        // Used by the history panel
-        $$.on(djDebug, "click", ".switchHistory", function (event) {
-            event.preventDefault();
-            const newStoreId = this.dataset.storeId;
-            const tbody = this.closest("tbody");
-
-            tbody
-                .querySelector(".djdt-highlighted")
-                .classList.remove("djdt-highlighted");
-            this.closest("tr").classList.add("djdt-highlighted");
-
-            ajaxForm(this).then(function (data) {
-                djDebug.setAttribute("data-store-id", newStoreId);
-                Object.keys(data).forEach(function (panelId) {
-                    if (djDebug.querySelector("#" + panelId)) {
-                        djDebug.querySelector("#" + panelId).outerHTML =
-                            data[panelId].content;
-                        djDebug.querySelector("#djdt-" + panelId).outerHTML =
-                            data[panelId].button;
-                    }
-                });
-            });
-        });
-
-        // Used by the history panel
-        $$.on(djDebug, "click", ".refreshHistory", function (event) {
-            event.preventDefault();
-            const container = djDebug.querySelector("#djdtHistoryRequests");
-            ajaxForm(this).then(function (data) {
-                if (data.requests.constructor === Array) {
-                    data.requests.forEach(function (request) {
-                        if (
-                            !container.querySelector(
-                                '[data-store-id="' + request.id + '"]'
-                            )
-                        ) {
-                            container.innerHTML =
-                                request.content + container.innerHTML;
-                        }
-                    });
-                }
             });
         });
